# openapi.yml
openapi: 3.0.3
info:
  title: API Bancaria - Ejercicio Técnico
  version: 1.0.0
  description: API para gestionar Clientes, Cuentas, Movimientos y generacion de reportes (JSON / PDF base64).
servers:
  - url: /api
    description: API base (prefijo /api en paths)
...
paths:
...
  /movimientos:
    get:
      summary: Listar movimientos
      operationId: listMovements
      tags:
        - Movements
      parameters:
        - name: clienteId
          in: query
          schema:
            type: string
            format: uuid
        - name: numeroCuenta
          in: query
          schema:
            type: string
        - name: fechaDesde
          in: query
          schema:
            type: string
            format: date
        - name: fechaHasta
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovimientoDto'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      summary: Crear movimiento (débito/crédito)
      operationId: createMovement
      tags:
        - Movements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoCreateDto'
      responses:
        '201':
          description: Movimiento creado y saldo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovimientoDto'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Conflicto de concurrencia (optimistic lock / version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /movimientos/{movimientoId}:
    get:
      summary: Obtener movimiento por ID
      operationId: getMovementById
      tags:
        - Movements
      parameters:
        - name: movimientoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Movimiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovimientoDto'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    put:
      summary: Actualizar movimiento (si aplica)
      operationId: updateMovement
      tags:
        - Movements
      parameters:
        - name: movimientoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoUpdateDto'
      responses:
        '200':
          description: Movimiento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovimientoDto'
    delete:
      summary: Eliminar (marcar inactivo) movimiento
      operationId: deleteMovement
      tags:
        - Movements
      parameters:
        - name: movimientoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Movimiento eliminado / marcado inactivo
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /reportes:
    get:
      summary: Generar estado de cuenta / reporte para un cliente en un rango de fechas
      operationId: generateReport
      tags:
        - Reports
      parameters:
        - name: clienteId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: fechaDesde
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaHasta
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: formato
          in: query
          description: "json (default) | pdf | pdfBase64"
          schema:
            type: string
            enum: [json, pdf, pdfBase64]
            default: json
      responses:
        '200':
          description: Reporte en JSON (estructura) o PDF
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ReportDto'
                  - $ref: '#/components/schemas/ReportWithPdfBase64Dto'
            application/pdf:
              schema:
                type: string
                format: binary
                description: "PDF binario; usar cuando formato=pdf y Accept: application/pdf"
        '404':
          description: Cliente o datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
components:
  schemas:
...
    MovimientoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cuentaId:
          type: string
          format: uuid
        fecha:
          type: string
          format: date-time
        tipo:
          type: string
          enum: ["DEBITO", "CREDITO"]
        valor:
          type: number
          format: decimal
          multipleOf: 0.01
        saldoPostMovimiento:
          type: number
          format: decimal
          multipleOf: 0.01
        createdAt:
          type: string
          format: date-time
      required: [id, cuentaId, fecha, tipo, valor, saldoPostMovimiento]
      example:
        id: "f1234567-89ab-4cde-0123-abcdefabcdef"
        cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
        fecha: "2025-08-28T10:30:00Z"
        tipo: "DEBITO"
        valor: 575.00
        saldoPostMovimiento: 1425.00
        createdAt: "2025-08-28T10:30:00Z"
    MovimientoCreateDto:
      type: object
      properties:
        cuentaId:
          type: string
          format: uuid
        tipo:
          type: string
          enum: ["DEBITO", "CREDITO"]
        valor:
          type: number
          format: decimal
          multipleOf: 0.01
      required: [cuentaId, tipo, valor]
      example:
        cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
        tipo: "DEBITO"
        valor: 575.00
    MovimientoUpdateDto:
      type: object
      properties:
        valor:
          type: number
          format: decimal
          multipleOf: 0.01
      example:
        valor: 600.00
    AccountSummaryDto:
      type: object
      properties:
        numeroCuenta:
          type: integer
          format: int64
        tipo:
          type: string
          enum: ["AHORROS", "CORRIENTE"]
        saldoInicial:
          type: number
          format: decimal
          multipleOf: 0.01
        saldoActual:
          type: number
          format: decimal
          multipleOf: 0.01
        totalDebitos:
          type: number
          format: decimal
          multipleOf: 0.01
        totalCreditos:
          type: number
          format: decimal
          multipleOf: 0.01
        movimientos:
          type: array
          items:
            $ref: '#/components/schemas/MovimientoDto'
      required: [numeroCuenta, tipo, saldoInicial, saldoActual, totalDebitos, totalCreditos, movimientos]
      example:
        numeroCuenta: 478758
        tipo: "AHORROS"
        saldoInicial: 2000.00
        saldoActual: 1425.00
        totalDebitos: 575.00
        totalCreditos: 0.00
        movimientos:
          - id: "f1234567-89ab-4cde-0123-abcdefabcdef"
            cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
            fecha: "2025-08-28T10:30:00Z"
            tipo: "DEBITO"
            valor: 575.00
            saldoPostMovimiento: 1425.00
            createdAt: "2025-08-28T10:30:00Z"
    ReportDto:
      type: object
      properties:
        clienteId:
          type: string
          format: uuid
        clienteNombre:
          type: string
        fechaDesde:
          type: string
          format: date
        fechaHasta:
          type: string
          format: date
        cuentas:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummaryDto'
        totalDebitos:
          type: number
          format: decimal
          multipleOf: 0.01
        totalCreditos:
          type: number
          format: decimal
          multipleOf: 0.01
      required: [clienteId, clienteNombre, fechaDesde, fechaHasta, cuentas, totalDebitos, totalCreditos]
      example:
        clienteId: "2d1f3b9a-e4aa-4fa8-9f9c-abcdef012345"
        clienteNombre: "Marianela Montalvo"
        fechaDesde: "2025-02-01"
        fechaHasta: "2025-02-28"
        totalDebitos: 575.00
        totalCreditos: 600.00
        cuentas:
          - numeroCuenta: 478758
            tipo: "AHORROS"
            saldoInicial: 2000.00
            saldoActual: 2025.00
            totalDebitos: 575.00
            totalCreditos: 600.00
            movimientos:
              - id: "f1234567-89ab-4cde-0123-abcdefabcdef"
                cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
                fecha: "2025-02-15T10:30:00Z"
                tipo: "DEBITO"
                valor: 575.00
                saldoPostMovimiento: 1425.00
                createdAt: "2025-02-15T10:30:00Z"
              - id: "g2345678-90bc-5def-1234-bcdefabcdefg"
                cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
                fecha: "2025-02-20T14:15:00Z"
                tipo: "CREDITO"
                valor: 600.00
                saldoPostMovimiento: 2025.00
                createdAt: "2025-02-20T14:15:00Z"
    ReportWithPdfBase64Dto:
      allOf:
        - $ref: '#/components/schemas/ReportDto'
        - type: object
          properties:
            pdfBase64:
              type: string
              format: byte
              description: "PDF codificado en base64 (cuando formato=pdfBase64)"
          required: [pdfBase64]
      example:
        clienteId: "2d1f3b9a-e4aa-4fa8-9f9c-abcdef012345"
        clienteNombre: "Marianela Montalvo"
        fechaDesde: "2025-02-01"
        fechaHasta: "2025-02-28"
        totalDebitos: 575.00
        totalCreditos: 600.00
        cuentas: []
        pdfBase64: "JVBERi0xLjQKJcOkw7zDtsO4w6DCjAoKMS..."
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
      required: [timestamp, status, error, message]
      example:
        timestamp: "2025-08-29T12:10:00Z"
        status: 400
        error: "Bad Request"
        code: "SALDO_NO_DISPONIBLE"
        message: "Saldo no disponible para la operacion solicitada"
        details:
          - "Cuenta 478758 tiene saldo insuficiente"
          - "Saldo actual: 100.00, monto solicitado: 575.00"